[#25] add swagger middleware & refactor global middlewares initialization
