+++ package.json
@@ -52,6 +52,7 @@
     "rimraf": "^3.0.2",
     "rxjs": "^7.5.5",
     "sqlite3": "^5.0.10",
+    "swagger-ui-express": "^4.5.0",
     "typeorm": "^0.3.7",
     "uuid": "^8.3.2"
   },
